Moving GrowthType to simulation_report.py
And other minor PR comments
