+++ covid_graphs/covid_graphs/simulation_report.py
@@ -1,3 +1,4 @@
+from enum import Enum
 from dataclasses import dataclass
 from pathlib import Path
 from typing import List
@@ -5,8 +6,16 @@ from typing import List
 from .pb.simulation_results_pb2 import SimulationResults
 
 
+class GrowthType(Enum):
+    Exponential = 'exponential'
+    Polynomial = 'polynomial'
+
+    def __str__(self):
+        return self.value
+
+
 @dataclass
-class SimulationReport():
+class SimulationReport:
     daily_positive: List[List[int]]
     daily_infected: List[List[int]]
     deltas: List[float]
@@ -16,8 +25,8 @@ class SimulationReport():
     # We want to store either alpha or gamma2. We could have two classes, one having alpha and the
     # other gamma2 (in C++ I would make a template). Don't know what's best practice in these
     # situations in Python.
-    param_name: str
     param: float
+    growth_type: GrowthType
 
 
 def create_simulation_reports(simulation_pb2_file: Path) -> List[SimulationReport]:
@@ -30,13 +39,13 @@ def create_simulation_reports(simulation_pb2_file: Path) -> List[SimulationRepor
         daily_positive = [run.daily_positive for run in result.runs]
         daily_infected = [run.daily_infected for run in result.runs]
         if result.HasField("alpha"):
-            param_name = "alpha"
-            best_param = result.alpha
+            growth_type = GrowthType.Polynomial
+            param = result.alpha
         else:
-            param_name = "gamma_2"
-            best_param = result.gamma2
+            growth_type = GrowthType.Exponential
+            param = result.gamma2
 
         reports.append(
             SimulationReport(daily_positive, daily_infected, result.deltas, result.b0,
-                             result.prefix_length, result.summary.error, param_name, best_param))
+                             result.prefix_length, result.summary.error, param, growth_type))
     return reports
